Added a decaying health effect for any entities with HP component

+++ include/HealthComponent.h
@@ -9,7 +9,7 @@ struct HealthComponent : public Component
 {
     HealthComponent() : Component("hp") {};
     void setName(std::string t_name) {id = t_name;};
-    int health{100};
+    float health{100.0f};
 };
 
 #endif
